refactor(be): update import paths
