+++ backend/src/auth/auth.module.ts
@@ -4,7 +4,7 @@ import { PassportModule } from '@nestjs/passport'
 import { UserModule } from 'src/user/user.module'
 import { AuthController } from './auth.controller'
 import { AuthService } from './auth.service'
-import { SECRET_KEY } from './jwt.constants'
+import { SECRET_KEY } from './config/jwt.config'
 import { JwtStrategy } from './strategy/jwt.strategy'
 import { LocalStrategy } from './strategy/local.strategy'
